Microsoft/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microsoft-query-processor"
version = "1.0.0"
description = "Production-ready query processor with multi-modal anomaly detection"
readme = "README.md"
authors = [
    {name = "Microsoft Advanced Query Processor Team", email = "example@microsoft.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "pyod>=1.0.0",
    "rich>=10.0.0",
    "psutil>=5.8.0",
    "ollama>=0.1.0",
]

[project.optional-dependencies]
azure = [
    "azure-storage-blob>=12.8.0",
    "azure-ai-anomalydetector>=3.0.0",
    "azure-identity>=1.6.0",
    "azure-monitor-ingestion>=1.0.0",
]
ms365 = [
    "msal>=1.16.0",
    "requests>=2.26.0",
]
production = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.2",
    "python-dotenv>=0.19.0"
]
all = [
    "azure-storage-blob>=12.8.0",
    "azure-ai-anomalydetector>=3.0.0",
    "azure-identity>=1.6.0",
    "azure-monitor-ingestion>=1.0.0",
    "msal>=1.16.0",
    "requests>=2.26.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.2",
    "python-dotenv>=0.19.0"
]

[project.scripts]
query-processor = "query_processor.main:main"

[project.urls]
"Homepage" = "https://github.com/importstring/microsoft-challenge-demo"
"Bug Tracker" = "https://github.com/importstring/microsoft-challenge-demo/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 